refactor(pokedex): Separation of concerns - gird render and view
